tests/wpt/web-platform-tests/css/css3-selectors/i18n/css3-selectors-lang-050.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>[lang="en-GB"], lang="en-US"</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-selectors/#attribute-selectors'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style type='text/css'>
.test div { width: 50px; }
#box[lang='en-GB'] { width: 100px; }
</style>
</head>
<body>



<div class="test"><div id="box" lang="en-US">&#xA0;</div></div>


<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 50);
}, "A lang= value and a lang attribute value will NOT match if their region subtags differ.");
</script>

<div id='log'></div>

</body>
</html>
